FIx compiler whines

+++ src/egtb/egtb_gen_dtm50.h
@@ -38,7 +38,7 @@
 // hmc=99 of each fusion runs the full chess-legality check per cell; every
 // lower-hmc layer piggybacks the ILLEGAL flag from opp[k+1] at the same idx.
 
-inline constexpr size_t DTM50_HMC_COUNT = 100;
+inline constexpr int DTM50_HMC_COUNT = 100;
 
 struct DTM50_Interrupted
 {
